Fixes for ARM architecture.
svn path=/trunk/koffice/; revision=1013597

+++ karbon/plugins/tools/KarbonCurveFit.cpp
@@ -54,7 +54,7 @@ class FitVector {
 
 	void normalize(){
 		qreal len=length();
-        if (qFuzzyCompare(len, 0.0)) {
+        if (qFuzzyCompare(len, qreal(0.0))) {
             return;
         }
 		m_X/=len; m_Y/=len;
@@ -67,7 +67,7 @@ class FitVector {
 
 	void scale(qreal s){
 		qreal len = length();
-        if (qFuzzyCompare(len, 0.0)) {
+        if (qFuzzyCompare(len, qreal(0.0))) {
             return;
         }
 		m_X *= s/len;
@@ -133,7 +133,7 @@ static qreal *ChordLengthParameterize(const QList<QPointF> &points,int first,int
     }
 
     qreal denominator = u[last-first];
-    if (qFuzzyCompare(denominator, 0.0)) {
+    if (qFuzzyCompare(denominator, qreal(0.0))) {
         denominator = Zero;
     }
     
@@ -284,9 +284,9 @@ QPointF* GenerateBezier(const QList<QPointF> &points, int first, int last, qreal
     det_X_C1  = X[0]    * C[1][1] - X[1]    * C[0][1];
 
     /* Finally, derive alpha values	*/
-    if (qFuzzyCompare(det_C0_C1, 0.0)) {
+    if (qFuzzyCompare(det_C0_C1, qreal(0.0))) {
 		det_C0_C1 = (C[0][0] * C[1][1]) * 10e-12;
-        if (qFuzzyCompare(det_C0_C1, 0.0)) {
+        if (qFuzzyCompare(det_C0_C1, qreal(0.0))) {
             det_C0_C1 = Zero;
         }
     }
@@ -421,7 +421,7 @@ static qreal NewtonRaphsonRootFind(QPointF *Q,QPointF P,qreal u)
     denominator = (Q1_u.x()) * (Q1_u.x()) + (Q1_u.y()) * (Q1_u.y()) +
 		      	  (Q_u.x() - P.x()) * (Q2_u.x()) + (Q_u.y() - P.y()) * (Q2_u.y());
     
-    if (qFuzzyCompare(denominator, 0.0)) {
+    if (qFuzzyCompare(denominator, qreal(0.0))) {
         denominator = Zero;
     }
